+++ chapter_16/Text_Listings/convertHex_C/hexToInt.h
@@ -1,7 +1,7 @@
-// Converts hex character string to int
+// Converts hex character string to long int
 // Returns number of characters
  
 #ifndef HEXTOINT_H
 #define HEXTOINT_H
-int hexToInt(char *stringPtr, int *intPtr);
+int hexToInt(char *stringPtr, long int *intPtr);
 #endif
